parser/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openbk==1.0.0
python-multipart==0.0.9
supabase==2.3.0
python-dotenv==1.0.0
pandas==2.1.4